add the return of register if user exists

+++ ComputationalThinkingPlatformApplication/src/main/java/com/example/model/LearningRecord.java
@@ -12,8 +12,8 @@ import javax.persistence.Id;
 public class LearningRecord {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private Long userId;
+    private int id;
+    private int userId;
     private String scene;
     private String progress;
 }
